fixes Errors when attempting to add `opentelemetry` to plugin BOM #1032
